android-13.0.0_r3 part 1

+++ external/mtools/buffer.h
@@ -20,9 +20,9 @@
 
 #include "stream.h"
 
-Stream_t *buf_init(Stream_t *Next, 
-		   int size, 
-		   int cylinderSize,
-		   int sectorSize);
+Stream_t *buf_init(Stream_t *Next,
+		   size_t size,
+		   size_t cylinderSize,
+		   size_t sectorSize);
 
 #endif
