+++ external/parameter-framework/upstream/test/test-platform/TestPlatform.h
@@ -144,7 +144,7 @@ private:
                                            std::string &strResult);
     bool createInclusiveSelectionCriterion(const std::string &strName, size_t nbValues,
                                            std::string &strResult);
-    bool setCriterionState(const std::string &strName, uint32_t uiState, std::string &strResult);
+    bool setCriterionState(const std::string &strName, uint64_t uiState, std::string &strResult);
     bool setCriterionStateByLexicalSpace(const IRemoteCommand &remoteCommand,
                                          std::string &strResult);
